+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Utils.java
@@ -109,6 +109,9 @@ abstract class Utils {
         } else if (value == QueryBuilder.BIND_MARKER) {
             sb.append("?");
             return true;
+        } else if (value == null) {
+            sb.append("null");
+            return true;
         } else {
             return false;
         }