+++ driver-core/CHANGELOG.rst
@@ -2,6 +2,7 @@ CHANGELOG
 =========
 
 * 1.0.1:
+  - [new] null values support in the Query builder.
   - [improvement] Better error message when no host are available.
   - [improvement] Improves performance of the stress example application been.